# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Alicerce WhatsBot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ConversationStatus = typeof ConversationStatus[keyof typeof ConversationStatus];


export const ConversationStatus = {
  bot: 'bot',
  human: 'human',
  closed: 'closed',
} as const;

export interface Conversation {
  id: number;
  whatsappNumber: string;
  /** @nullable */
  customerName?: string | null;
  status: ConversationStatus;
  /** @nullable */
  lastMessage?: string | null;
  unreadCount?: number;
  createdAt: string;
  updatedAt: string;
}

export interface DashboardStats {
  totalConversations: number;
  activeBot: number;
  awaitingHuman: number;
  totalProducts: number;
  pendingOrders: number;
  pendingBudgets: number;
  recentConversations: Conversation[];
}

export interface Product {
  id: number;
  name: string;
  category: string;
  price: number;
  unit: string;
  /** @nullable */
  description?: string | null;
  inStock: boolean;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  category: string;
  price: number;
  unit: string;
  description?: string;
  inStock?: boolean;
}

export type MessageDirection = typeof MessageDirection[keyof typeof MessageDirection];


export const MessageDirection = {
  inbound: 'inbound',
  outbound: 'outbound',
} as const;

export interface Message {
  id: number;
  conversationId: number;
  content: string;
  direction: MessageDirection;
  messageType?: string;
  createdAt: string;
}

export type ConversationDetailStatus = typeof ConversationDetailStatus[keyof typeof ConversationDetailStatus];


export const ConversationDetailStatus = {
  bot: 'bot',
  human: 'human',
  closed: 'closed',
} as const;

export interface ConversationDetail {
  id: number;
  whatsappNumber: string;
  /** @nullable */
  customerName?: string | null;
  status: ConversationDetailStatus;
  /** @nullable */
  lastMessage?: string | null;
  unreadCount?: number;
  messages: Message[];
  createdAt: string;
  updatedAt: string;
}

export interface MessageInput {
  content: string;
}

export type ConversationStatusUpdateStatus = typeof ConversationStatusUpdateStatus[keyof typeof ConversationStatusUpdateStatus];


export const ConversationStatusUpdateStatus = {
  bot: 'bot',
  human: 'human',
  closed: 'closed',
} as const;

export interface ConversationStatusUpdate {
  status: ConversationStatusUpdateStatus;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderItem {
  id: number;
  /** @nullable */
  productId?: number | null;
  productName: string;
  quantity: number;
  unitPrice: number;
}

export interface Order {
  id: number;
  whatsappNumber: string;
  /** @nullable */
  customerName?: string | null;
  status: OrderStatus;
  total: number;
  /** @nullable */
  notes?: string | null;
  items?: OrderItem[];
  createdAt: string;
  updatedAt: string;
}

export interface OrderItemInput {
  productId?: number;
  productName: string;
  quantity: number;
  unitPrice: number;
}

export interface OrderInput {
  whatsappNumber: string;
  customerName?: string;
  notes?: string;
  items: OrderItemInput[];
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export type BudgetRequestStatus = typeof BudgetRequestStatus[keyof typeof BudgetRequestStatus];


export const BudgetRequestStatus = {
  pending: 'pending',
  attended: 'attended',
  closed: 'closed',
} as const;

export interface BudgetRequest {
  id: number;
  whatsappNumber: string;
  /** @nullable */
  customerName?: string | null;
  description: string;
  status: BudgetRequestStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type BudgetRequestUpdateStatus = typeof BudgetRequestUpdateStatus[keyof typeof BudgetRequestUpdateStatus];


export const BudgetRequestUpdateStatus = {
  pending: 'pending',
  attended: 'attended',
  closed: 'closed',
} as const;

export interface BudgetRequestUpdate {
  status?: BudgetRequestUpdateStatus;
  notes?: string;
}

export interface StoreSettings {
  id: number;
  storeName: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  whatsappNumber?: string | null;
  /** @nullable */
  openingHours?: string | null;
  /** @nullable */
  evolutionApiUrl?: string | null;
  /** @nullable */
  evolutionApiKey?: string | null;
  /** @nullable */
  evolutionInstance?: string | null;
  /** @nullable */
  botWelcomeMessage?: string | null;
}

export interface StoreSettingsInput {
  storeName?: string;
  address?: string;
  phone?: string;
  whatsappNumber?: string;
  openingHours?: string;
  evolutionApiUrl?: string;
  evolutionApiKey?: string;
  evolutionInstance?: string;
  botWelcomeMessage?: string;
}

export interface EvolutionPayload { [key: string]: unknown }

export interface WebhookResult {
  ok: boolean;
  message?: string;
}

export type ListProductsParams = {
category?: string;
inStock?: boolean;
};

export type ListConversationsParams = {
status?: ListConversationsStatus;
};

export type ListConversationsStatus = typeof ListConversationsStatus[keyof typeof ListConversationsStatus];


export const ListConversationsStatus = {
  bot: 'bot',
  human: 'human',
  closed: 'closed',
} as const;

export type ListOrdersParams = {
status?: ListOrdersStatus;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export type ListBudgetRequestsParams = {
status?: ListBudgetRequestsStatus;
};

export type ListBudgetRequestsStatus = typeof ListBudgetRequestsStatus[keyof typeof ListBudgetRequestsStatus];


export const ListBudgetRequestsStatus = {
  pending: 'pending',
  attended: 'attended',
  closed: 'closed',
} as const;