# lib/api-zod/src/generated/types/budgetRequest.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Alicerce WhatsBot API
 * OpenAPI spec version: 0.1.0
 */
import type { BudgetRequestStatus } from './budgetRequestStatus';

export interface BudgetRequest {
  id: number;
  whatsappNumber: string;
  /** @nullable */
  customerName?: string | null;
  description: string;
  status: BudgetRequestStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}