# lib/api-zod/src/generated/types/budgetRequestUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Alicerce WhatsBot API
 * OpenAPI spec version: 0.1.0
 */
import type { BudgetRequestUpdateStatus } from './budgetRequestUpdateStatus';

export interface BudgetRequestUpdate {
  status?: BudgetRequestUpdateStatus;
  notes?: string;
}