# lib/api-zod/src/generated/types/conversationDetail.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Alicerce WhatsBot API
 * OpenAPI spec version: 0.1.0
 */
import type { ConversationDetailStatus } from './conversationDetailStatus';
import type { Message } from './message';

export interface ConversationDetail {
  id: number;
  whatsappNumber: string;
  /** @nullable */
  customerName?: string | null;
  status: ConversationDetailStatus;
  /** @nullable */
  lastMessage?: string | null;
  unreadCount?: number;
  messages: Message[];
  createdAt: Date;
  updatedAt: Date;
}