# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Alicerce WhatsBot API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  whatsappNumber: string;
  /** @nullable */
  customerName?: string | null;
  status: OrderStatus;
  total: number;
  /** @nullable */
  notes?: string | null;
  items?: OrderItem[];
  createdAt: Date;
  updatedAt: Date;
}