# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Alicerce WhatsBot API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemInput } from './orderItemInput';

export interface OrderInput {
  whatsappNumber: string;
  customerName?: string;
  notes?: string;
  items: OrderItemInput[];
}