# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Alicerce WhatsBot API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  category: string;
  price: number;
  unit: string;
  /** @nullable */
  description?: string | null;
  inStock: boolean;
  createdAt: Date;
}