# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Alicerce WhatsBot API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  category: string;
  price: number;
  unit: string;
  description?: string;
  inStock?: boolean;
}