# lib/api-zod/src/generated/types/webhookResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Alicerce WhatsBot API
 * OpenAPI spec version: 0.1.0
 */

export interface WebhookResult {
  ok: boolean;
  message?: string;
}